Fix Linux build - replace INT64 with long long
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@2214 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Stylization/RS_Raster.h
@@ -44,7 +44,7 @@ public:
     virtual RS_String       GetVerticalUnits()  = 0;
 
     virtual int             GetNullValueType()  = 0;
-    virtual INT64           GetNullValueData()  = 0;
+    virtual long long       GetNullValueData()  = 0;
 };
 
 #endif
